+++ example/example_task_cancel_group_by_failure.dart
@@ -9,14 +9,14 @@ Future<void> main() async {
   final group = <Task<int>>[];
 
   void onExit(AnyTask task) {
-    if (!task.isCompleted) {
+    if (!task.isSuccessful) {
       cts.cancel();
     }
   }
 
   parent = Task.run<void>(name: 'Parent', () async {
     Task.onExit((task) {
-      print('On exit: ${task.toString()} (${task.state.name})');
+      print('On exit: ${task.toString()} (${task.status.name})');
       onExit(task);
     });
 
@@ -24,7 +24,7 @@ Future<void> main() async {
     for (var i = 1; i <= 3; i++) {
       final t = Task<int>(name: 'Child $i', () async {
         Task.onExit((task) {
-          print('On exit: ${task.toString()} (${task.state.name})');
+          print('On exit: ${task.toString()} (${task.status.name})');
           onExit(task);
         });